Separate DimensionOrderDropdown and add tooltip
